Refactoring Templates Implementation
