+++ DoublyLinkedList.h
@@ -1,48 +1,79 @@
-
 #ifndef LRUPROJECT_DOUBLYLINKEDLIST_H
 #define LRUPROJECT_DOUBLYLINKEDLIST_H
 
-#include <iostream>
+#include "Node.h"
 
-class Node {
-public:
-    char key;
-    int value;
-    Node* prev;
-    Node* next;
-    // Default Constructor
-    Node() {
-        prev = nullptr;
-        next = nullptr;
-        key = 'a';
-        value = -1;
-    }
-    // Custom Constructor
-    Node(char key, int value){
-        this->key = key;
-        this->value = value;
-        prev = nullptr;
-        next = nullptr;
-    }
+/**
+ * Doubly Linked List - Definition
+ * Self Organized List
+ */
 
-    void killMe(){
-        std::cout << "** killMe **\n";
-        if(this->next){
-            this->next->killMe();
-        }
-        std::cout << "Delete this Node: (" << key << ": " << value << ") \n";
-        delete this;
-    }
-};
-
+template<class T1, class T2>
 class DoublyLinkedList {
 public:
-    Node* head;     // most recently used
-    Node* tail;     // least recently used
+    Node<T1, T2>* head;     // most recently used
+    Node<T1, T2>* tail;     // least recently used
     DoublyLinkedList();
-    void push_front(Node* node);
+    void push_front(Node<T1, T2>* node);
     void print();
     void printReverse();
 };
+
+/**
+ * Implementation Methods
+ */
+template<class T1, class T2>
+DoublyLinkedList<T1, T2>::DoublyLinkedList() {
+    head = nullptr;
+    tail = nullptr;
+}
+
+template<class T1, class T2>
+void DoublyLinkedList<T1, T2>::push_front(Node<T1, T2> *node) {
+    if(head){
+        Node<T1, T2>* cpHead = head;
+        head->prev = node;
+        head = node;
+        node->next = cpHead;
+    }
+    else {  // if does not exist head
+        head = node;
+        tail = node;
+    }
+}
+
+template<class T1, class T2>
+void DoublyLinkedList<T1, T2>::print() {
+    std::cout << "** Printing Doubly Linked List **\n";
+    Node<T1, T2>* temp = head;
+    while(temp){
+        if(temp->next){
+            std::cout << "( " << temp->key << ": " << temp->value << " ) <--> ";
+        }
+        else{
+            std::cout << "( " << temp->key << ": " << temp->value << " ) --> ";
+        }
+        temp = temp->next;
+    }
+    std::cout << "\n";
+}
+
+template<class T1, class T2>
+void DoublyLinkedList<T1, T2>::printReverse() {
+    std::cout << "** Printing in 'Reverse' the Doubly Linked List **\n";
+    Node<T1, T2>* temp = tail;
+    std::cout << "Tail...";
+    while(temp){
+        if(temp->prev){
+            std::cout << "( " << temp->key << ": " << temp->value << " ) <--> ";
+        }
+        else{
+            std::cout << "( " << temp->key << ": " << temp->value << " ) ";
+        }
+        temp = temp->prev;
+    }
+    std::cout << "...Head\n";
+}
+
 
 #endif //LRUPROJECT_DOUBLYLINKEDLIST_H
